tests/wpt/tests/encoding-detection/ja-ISO-2022-JP-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>ja ISO-2022-JP late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "windows-1252", 'Expected windows-1252 due to ISO-2022-JP no longer getting detected');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/ja-ISO-2022-JP-late.sub.html");
</script>
